+++ changelog.md
@@ -24,6 +24,7 @@ Codion Change Log
 - Utilities.linkToEnabledState() renamed enableActions() and enableComponents().
 - ValidIndicatorFactory.instance() now returns Optional.
 - AbstractComponentBuilder no longer throws exception if the specified ValidIndicatorFactory is not found.
+- ValidIndicatorFactory.instance() overloaded with factoryClassName parameter.
 ### is.codion.swing.framework.ui
 - EntityApplicationPanel.Builder.applicationModelFactory() and applicationPanelFactory() renamed applicationModel() and applicationPanel().
 - EntityApplicationPanel.createLogControls() now protected final.